Added the inoutTypes.go that have all the structures
